--- src/constants.rs
@@ -1,30 +1,0 @@
-// 
-pub const STARTING_WOOD: i32 = 500;
-pub const BUILDER_COST: i32 = 3;
-pub const WORKER_COST: i32 = 3;
-
-
-// Map 
-pub const SMOOTHNESS: f32 = 0.75; //0.75 is a good one
-pub const TILE_SIZE:f32 = 50.0; //50
-pub const MAP_WIDTH:i32 = 25;
-pub const MAP_HEIGHT:i32 = 25;
-
-
-// Bandit
-pub const BANDIT_ENEMY_RANGE: f32 = 500.0;
-pub const BANDIT_ATTACK_RANGE: f32 = 75.0;
-
-// WARRIOR
-pub const WARRIOR_ENEMY_RANGE: f32 = 500.0;
-pub const WARRIOR_ATTACK_RANGE: f32 = 75.0;
-
-// Combat
-pub const KNOCKBACK: f32 = 1000.0;
-
-// Window
-pub const WINDOW_HEIGHT: f32 = 1600.0;
-pub const WINDOW_WIDTH: f32 = 1200.0;
-
-// Camera
-pub const CAMERA_ZOOM: f32 = 2.0;
