+++ src/entities/mod.rs
@@ -1,2 +1,7 @@
 pub mod bandit;
 pub mod builder;
+pub mod minion;
+pub mod rabbit;
+pub mod hunter;
+pub mod plugin;
+pub mod peasant;
