+++ src/game/gold/mod.rs
