+++ src/game/mod.rs
@@ -1,0 +1,8 @@
+pub mod gold;
+pub mod camp;
+pub mod campsite;
+pub mod tree;
+pub mod plugin;
+pub mod movement;
+pub mod map;
+pub mod camera;
