+++ src/lib.rs
@@ -1,41 +1,22 @@
 pub mod startup;
 pub mod player;
 pub mod window_icon;
-pub mod map;
-pub mod minion;
-//pub mod builder;
-pub mod campsite;
 pub mod camera;
-pub mod tree;
-//pub mod bandit;
 pub mod utility;
 pub mod ui;
 pub mod combat;
-pub mod movement; 
-pub mod gold;
-pub mod peasant;
-pub mod camp;
-pub mod constants;
 pub mod common;
 pub mod entities;
+pub mod debug;
+pub mod game;
 
-
+pub use game::*;
+pub use debug::*;
 pub use entities::*;
 pub use common::*;
-pub use constants::*;
-pub use camp::*;
-pub use peasant::*;
-pub use gold::*;
-pub use movement::*;
 pub use combat::*; 
 pub use ui::*;
 pub use utility::*;
-//pub use bandit::*;
-pub use tree::*;
-pub use campsite::*;
 pub use player::*;
-pub use minion::*;
-pub use map::*;
 pub use startup::*;
-//pub use builder::*;
 pub use camera::*;
